añadiendo cambios a los menus
